Examenes/ExamenSegundoParcial/Vehiculos/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
python-dotenv
PyJWT
grpcio
grpcio-tools